Solution built changed to .Net 4 instead of mono. - Adding a Lib directory containing the missing frameworks - Changing some method signatures and namespaces that are different in .Net - Some encrypt libraries are missing due that there is no direct port in   .Net (not a priority at this momment) - There are failing tests pending review

+++ Sharpen/Sharpen/Extensions.cs
@@ -604,10 +604,10 @@ namespace Sharpen
 			return val;
 		}
 
-		public static int GetTotalInFixed (this Inflater inf)
+		public static long GetTotalInFixed (this Inflater inf)
 		{
 			if (inf.TotalIn > 0)
-				return inf.TotalIn + 4;
+				return inf.TotalIn + (long) 4;
 			else
 				return 0;
 		}
